+++ src/coordinax/_src/vectors/transform/dn.py
@@ -29,23 +29,23 @@ def vconvert(target: type[Space], w: PoincarePolarVector, /) -> Space:
     PoincarePolarVector(
         rho=Quantity[...](value=f32[1,2], unit=Unit("m")),
         pp_phi=Quantity[...]( value=f32[1,2], unit=Unit("m rad(1/2) / s(1/2)") ),
-        z=Quantity[...](value=f32[1,2], unit=Unit("m")),
-        d_rho=Quantity[...]( value=f32[1,2], unit=Unit("m / s") ),
-        d_pp_phi=Quantity[...]( value=f32[1,2], unit=Unit("m rad(1/2) / s(1/2)") ),
-        d_z=Quantity[...]( value=f32[1,2], unit=Unit("m / s") )
+      z=Quantity[...](value=i32[1,2], unit=Unit("m")),
+      d_rho=Quantity[...]( value=f32[1,2], unit=Unit("m / s") ),
+      d_pp_phi=Quantity[...]( value=f32[1,2], unit=Unit("m rad(1/2) / s(1/2)") ),
+      d_z=Quantity[...]( value=f32[1,2], unit=Unit("m / s") )
     )
 
     >>> cx.vconvert(cx.Space, w)
     Space({
         'length': CartesianPos3D(
-            x=Quantity[...](value=f32[1,2], unit=Unit("m")),
-            y=Quantity[...](value=f32[1,2], unit=Unit("m")),
-            z=Quantity[...](value=f32[1,2], unit=Unit("m"))
+            x=Quantity[...](value=i32[1,2], unit=Unit("m")),
+            y=Quantity[...](value=i32[1,2], unit=Unit("m")),
+            z=Quantity[...](value=i32[1,2], unit=Unit("m"))
         ),
         'speed': CartesianVel3D(
-            d_x=Quantity[...]( value=f32[1,2], unit=Unit("m / s") ),
-            d_y=Quantity[...]( value=f32[1,2], unit=Unit("m / s") ),
-            d_z=Quantity[...]( value=f32[1,2], unit=Unit("m / s") )
+            d_x=Quantity[...]( value=i32[1,2], unit=Unit("m / s") ),
+            d_y=Quantity[...]( value=i32[1,2], unit=Unit("m / s") ),
+            d_z=Quantity[...]( value=i32[1,2], unit=Unit("m / s") )
         )
     })
